pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sf-common"
version = "0.1.0"
description = "SemanticForge common utilities: config, exceptions, logging, observability, models, utils"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [{ name = "SemanticForge Team" }]
dependencies = [
  "pydantic==2.5.0",
  "python-dotenv>=1.0,<2.0",
  "PyYAML>=6.0,<7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
  "common",
  "common.config",
  "common.exceptions",
  "common.logging",
  "common.models",
  "common.observability",
  "common.utils",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

